Fix printf/snprintf related warnings
- cl_generic.c: Change pos from int to size_t for snprintf buffer
  calculations, avoiding sign conversion warnings
- violalex.c: Use %ld format for long integer (yylval.i)
- mcidas.c: Cast dir.satid to int when calling mc_sensor()

+++ src/libIMG/mcidas.c
@@ -107,7 +107,7 @@ Image* mcidasLoad(char* fullname, char* name, unsigned int verbose)
     }
 
     if (verbose)
-        (void)printf("%s %lu %lu (%lu, %lu) (%lu, %lu)\n", mc_sensor(dir.satid), dir.idate, dir.itime,
+        (void)printf("%s %lu %lu (%lu, %lu) (%lu, %lu)\n", mc_sensor((int)dir.satid), dir.idate, dir.itime,
                      dir.lcor, dir.ecor, dir.lres, dir.eres);
     znocache(zf);
     /* skip the nav */
